pyproject.toml
[tool.poetry]
name = "syndata"
version = "0.1.0"
description = "Generator for realistic, but not real data für eHealth Applications"
authors = ["spilikin <108782+spilikin@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
icecream = "^2.1.3"
uvicorn = "^0.24.0.post1"
fhir-resources = "^7.0.2"
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
